requirements.txt
BeautifulSoup==3.2.1
Django==1.5.1
Fabric==1.7.0
MySQL-python==1.2.4
PIL==1.1.7
South==0.8.2
boto==2.9.9
dj-database-url==0.2.2
dj-static==0.0.5
django-debug-toolbar==0.9.4
django-picklefield==0.3.0
django-pipeline==1.3.14
django-pylibmc-sasl==0.2.4
django-storages==1.1.8
django-toolbelt==0.0.1
futures==2.1.4
gunicorn==17.5
paramiko==1.11.0
pycrypto==2.6
pylibmc==1.2.3
raven==3.4.1
requests==1.2.3
six==1.3.0
static==0.4
wsgiref==0.1.2
